[test] move string helper functions into test library
