+++ src/test/settings_tests.cpp
@@ -4,8 +4,9 @@
 
 #include <util/settings.h>
 
-#include <test/util.h>
 #include <test/util/setup_common.h>
+#include <test/util/str.h>
+
 
 #include <boost/test/unit_test.hpp>
 #include <univalue.h>
